requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
alembic==1.15.2
pydantic[email]==2.11.3
pydantic-settings==2.9.1
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.20
aiofiles==24.1.0
httpx==0.28.1
aiobotocore==2.21.1
edge-tts==7.0.2
deep-translator==1.11.4
psycopg2-binary==2.9.10
